SessionService replaced with HttpSession, + more
The SessionService class (implemented as a Singleton) was being used to
track whether a used had logged in, and if the case of a logged in user,
which user was logged in.
This has now been replaced by the Java Spring HttpSession interface, which
should be a much more convenient way of handling this.
Some added functionality, mostly through the Java Persistance API, to list
all the Groups that a particular User is in.